cloud backend/requirements.txt
flask
flask-cors
python-whois
